Fix metricId's when using multiple vCenter (versions)
When using multiple VCenter (versions), not all the versions have the
same metricId for each metric.
As the Metrics config was passed by pointer every time, it was
overwritten by the next vcenter init.

This caused vcenter collection to fail, because the wrong metricId was
used.

Make sure the values are copied once, so they don't get overwritten.